chore: add interface across mesh shape for shape generation API
